rename: types/type.ts -> types/index.ts로 수정
