+++ src/components/create/question/scheduleQuestion.tsx
@@ -1,5 +1,5 @@
 import isModeDisabled from "@/lib/utils/isModeDisabled";
-import type { QuestionProps } from "@/types/types";
+import type { QuestionProps } from "@/types";
 import Options from "../options";
 
 const ScheduleQuestion: React.FC<QuestionProps> = ({ question, mode }) => {
