+++ src/components/form/formCardItem.tsx
@@ -3,7 +3,7 @@
 import { useEncryptedEntryNavigation } from "@/hooks/useEncryptedEntryNavigation";
 import formateDate from "@/lib/utils/formateDate";
 import getColorById from "@/lib/utils/getRandomColor";
-import type { Form } from "@/types/types";
+import type { Form } from "@/types";
 import Image from "next/image";
 import { useMemo } from "react";
 import Reaction from "./reaction";
