+++ src/constants/states.ts
@@ -1,4 +1,4 @@
-import type { ActionState } from "@/types/types";
+import type { ActionState } from "@/types";
 
 export const INITIAL_ACTION_STATE: ActionState = Object.freeze({
   status: null,
