+++ src/lib/firebase/user/fetchUserDataServer.ts
@@ -1,7 +1,7 @@
 // users docs에서 선택적으로 데이터 조합하여 반환
 
 import { getServerUid } from "@/lib/firebase/auth/getServerUid";
-import type { Form } from "@/types/types";
+import type { Form } from "@/types";
 import type {
   User,
   UserActivityFields,
